// libs/src/sdk/types/types.ts
export type Nullable<T> = T | null

type CID = string
type ID = number
// ...
export interface TrackSegment {
  duration: string
  multihash: CID
}
export interface Download {
  is_downloadable: Nullable<boolean>
  requires_follow: Nullable<boolean>
  cid: Nullable<string>
}

export type TokenStandard = 'ERC721' | 'ERC1155'

export type PremiumConditionsEthNFTCollection = {
  chain: 'eth'
  standard: TokenStandard
  address: string
  name: string
  slug: string
  imageUrl: Nullable<string>
  externalLink: Nullable<string>
}

export type PremiumConditionsSolNFTCollection = {
  chain: 'sol'
  address: string
  name: string
  imageUrl: Nullable<string>
  externalLink: Nullable<string>
}

export type PremiumConditions = {
  nft_collection?:
    | PremiumConditionsEthNFTCollection
    | PremiumConditionsSolNFTCollection
  follow_user_id?: number
  tip_user_id?: number
}

export type PremiumContentSignature = {
  data: string
  signature: string
}

export type TrackMetadata = {
  blocknumber: number
  activity_timestamp?: string
  is_delete: boolean
  track_id: number
  track_cid: string
  created_at: string
  isrc: Nullable<string>
  iswc: Nullable<string>
  credits_splits: Nullable<string>
  description: Nullable<string>
  download: Nullable<Download>
  genre: string
  has_current_user_reposted: boolean
  has_current_user_saved: boolean
  license: Nullable<string>
  mood: Nullable<string>
  play_count: number
  owner_id: ID
  release_date: Nullable<string>
  repost_count: number
  save_count: number
  tags: Nullable<string>
  title: string
  track_segments: TrackSegment[]
  cover_art: Nullable<CID>
  cover_art_sizes: Nullable<CID>
  is_unlisted: boolean
  is_available: boolean
  is_premium: boolean
  premium_conditions: Nullable<PremiumConditions>
  premium_content_signature: Nullable<PremiumContentSignature>
  listenCount?: number
  permalink: string

  // Optional Fields
  is_invalid?: boolean
  stem_of?: {
    parent_track_id: ID
  }

  // Added fields
  dateListened?: string
  duration: number

  is_playlist_upload?: boolean
}